sdk upgrade: workaround failing tests
The linter reported errors for a number of places (in existing code)
where the return value of calls to
controllerutil.SetControllerReference were not being checked. After
checks were added, the unit tests started failing. This indicates an
existing bug in the code. This change is to update the code so that it
will log the error and continue, so the build will be successful. An
issue will be raised for a proper fix for this.

Signed-off-by: John Pitman <jpitman@redhat.com>